+++ app/src/hooks/api/useProjectApi.ts
@@ -178,19 +178,28 @@ const useProjectApi = (axios: AxiosInstance) => {
   };
 
   /**
-   * Toggle visibility state of project attachments.
+   * Make visibility of project attachment private.
+   *
+   * @param {number} projectId
+   * @param {number} attachmentId
+   * @return {*}  {Promise<any>}
+   */
+  const makeAttachmentPrivate = async (projectId: number, attachmentId: number): Promise<any> => {
+    const { data } = await axios.put(`/api/project/${projectId}/attachments/${attachmentId}/makePrivate`);
+
+    return data;
+  };
+
+  /**
+   * Make visibility of project attachment public.
    *
    * @param {number} projectId
    * @param {number} attachmentId
    * @param {any} securityToken
    * @return {*}  {Promise<any>}
    */
-  const toggleProjectAttachmentVisibility = async (
-    projectId: number,
-    attachmentId: number,
-    securityToken: any
-  ): Promise<any> => {
-    const { data } = await axios.put(`/api/project/${projectId}/attachments/${attachmentId}/toggleVisibility`, {
+  const makeAttachmentPublic = async (projectId: number, attachmentId: number, securityToken: any): Promise<any> => {
+    const { data } = await axios.put(`/api/project/${projectId}/attachments/${attachmentId}/makePublic`, {
       securityToken
     });
 
@@ -249,7 +258,8 @@ const useProjectApi = (axios: AxiosInstance) => {
     deleteProject,
     publishProject,
     getPublicProjectsList,
-    toggleProjectAttachmentVisibility
+    makeAttachmentPublic,
+    makeAttachmentPrivate
   };
 };
